+++ web/client/test-resources/widgets/dependenciesToFiltersData.js
@@ -99,6 +99,60 @@ export const inputLayerFilterSTATENAME = {
     hits: false
 };
 
+export const layerFilter = {
+    "layerFilter": {
+        "searchUrl": null,
+        "featureTypeConfigUrl": null,
+        "showGeneratedFilter": false,
+        "attributePanelExpanded": true,
+        "spatialPanelExpanded": true,
+        "crossLayerExpanded": true,
+        "showDetailsPanel": false,
+        "groupLevels": 5,
+        "useMapProjection": false,
+        "toolbarEnabled": true,
+        "groupFields": [
+            {
+                "id": 1,
+                "logic": "OR",
+                "index": 0
+            }
+        ],
+        "maxFeaturesWPS": 5,
+        "filterFields": [
+            {
+                "rowId": 1613414722261,
+                "groupId": 1,
+                "attribute": "STATE_NAME",
+                "operator": "=",
+                "value": "Arizona",
+                "type": "string",
+                "fieldOptions": {
+                    "valuesCount": 0,
+                    "currentPage": 1
+                },
+                "exception": null,
+                "loading": false,
+                "openAutocompleteMenu": false,
+                "options": {
+                    "STATE_NAME": []
+                }
+            }
+        ],
+        "spatialField": {
+            "method": null,
+            "operation": "INTERSECTS",
+            "geometry": null,
+            "attribute": "the_geom"
+        },
+        "simpleFilterFields": [],
+        "crossLayerFilter": null,
+        "autocompleteEnabled": true,
+        "disabled": false
+    }
+};
+
+export const emptyLayerFilter = {}
 /**
  * output data for filters
  */
@@ -145,3 +199,4 @@ export const resultSpatialFilterMultiple =
     + "</ogc:Intersects>"
     + "</ogc:Or></ogc:And></ogc:Filter>";
 
+export const resultLayerFilter = `<ogc:Filter><ogc:And><ogc:Or><ogc:PropertyIsEqualTo><ogc:PropertyName>STATE_NAME</ogc:PropertyName><ogc:Literal>Arizona</ogc:Literal></ogc:PropertyIsEqualTo></ogc:Or></ogc:And></ogc:Filter>`;
